fix(data center compare panel): rename component, remove unnecessary code and add BEM styles
