+++ src/scss/layout/_data-center-compare-panel.scss
@@ -1,0 +1,123 @@
+.data-center-compare-panel {
+  position: absolute;
+  display: flex;
+  flex-direction: column;
+  padding-top: 5rem;
+  width: calc(14vw + 16rem);
+  background-color: $lighter-gray;
+  box-shadow: 0 24px 15px rgba(0,0,0,0.2);
+  box-sizing: border-box;
+  transition: transform 0.4s, box-shadow 0.4s;
+  z-index: 2;
+
+  // .data-center-compare-panel__box:not(:last-of-type)
+  &__box:not(:last-of-type) {
+    margin-bottom: 1rem;
+  }
+
+  // .data-center-compare-panel__button
+  &__button {
+    width: 200px;
+    margin-top: 2rem;
+    margin-left: auto;
+    margin-right: auto;
+    padding: 1rem;
+    background-color: #fff;
+    border-radius: 18px;
+    border: none;
+    box-shadow: 0 10px 15px rgba(0,0,0,0.1);
+    font-family: 'Euclid Triangle', 'Open Sans', sans-serif;
+
+    // .data-center-compare-panel__button:not(:disabled):hover
+    &:not(:disabled):hover {
+      cursor: pointer;
+    }
+  }
+
+  // .data-center-compare-panel__close
+  &__close {
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+    height: 3rem;
+    width: 3rem;
+    color: black;
+    border-radius: 4px;
+    background-color: #fff;
+    outline: 0;
+  }
+
+  // .data-center-compare-panel__collapsed
+  &__collapsed {
+    transform: translate(-100%);
+    box-shadow: 0 24px 15px rgba(0,0,0,0);
+  }
+
+  // .data-center-compare-panel__content
+  &__content {
+    overflow-x: visible;
+    overflow-y: auto;
+    padding: 0 3rem 3rem;
+    max-height: 75%;
+  }
+
+  // .data-center-compare-panel__data-center-box
+  &__data-center-box {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    padding: 1rem;
+    height: auto;
+    background-color: #fff;
+    border-radius: 0.3rem;
+    box-shadow: 0 10px 15px rgba(0,0,0,0.1);
+  }
+
+  // .data-center-compare-panel__headline
+  &__headline {
+    font-size: 0.85em;
+    font-weight: 700;
+    line-height: 1.25em;
+    margin-bottom: auto;
+  }
+
+  // .data-center-compare-panel__headline__remove-data-center
+  &__remove-data-center {
+    cursor: pointer;
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+    color: black;
+    outline: 0;
+  }
+
+  // .data-center-compare-panel__row:not(:last-of-type)
+  &__row:not(:last-of-type) {
+    margin-bottom: 1rem;
+  }
+
+  // .data-center-compare-panel__subtext
+  &__subtext {
+    margin-top: 0.375rem;
+    font-size: 0.85em;
+  }
+
+   // .data-center-compare-panel svg
+  & > svg {
+    display: block;
+  }
+
+  @include respond-to('medium-up') {
+    height: 100%; /* Should only be set on desktop */
+  }
+
+  @include respond-to('small') {
+    padding-bottom: 0;
+    position: relative;
+    width: 100%;
+  }
+}
