--- src/scss/layout/_data-center-compare.scss
@@ -1,111 +1,0 @@
-.panel {
-  background-color: $light-gray;
-  color: black;
-  font-size: 0.8rem;
-}
-
-/* Data-center collapse button */
-#data-center-compare-collapse-button {
-  right: 1rem;
-  top: 1rem;
-  position: absolute;
-  cursor: pointer;
-  border-radius: 4px;
-  background-color: $lighter-gray;
-  width: 3rem;
-  height: 3rem;
-  color: black;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  outline: 0;
-}
-
-.data-center-left-panel {
-  color: $black;
-  width: calc(14vw + 16rem);
-  background-color: $lighter-gray;
-  position: absolute;
-  box-shadow: 0 24px 15px rgba(0,0,0,0.2);
-  z-index: 2;
-  display: flex;
-  padding-top: 5rem;
-  flex-direction: column;
-  box-sizing: border-box;
-  transition: transform 0.4s, box-shadow 0.4s;
-
-  & .content {
-    overflow-x: visible;
-    overflow-y: auto;
-    padding: 0 3rem 3rem;
-    max-height: 75%;
-  }
-
-  & .box {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    padding: 1rem;
-    height: auto;
-    background-color: #fff;
-    border-radius: 1rem;
-    box-shadow: 0 10px 15px rgba(0,0,0,0.1);
-  }
-
-  & .box:not(:last-of-type) {
-    margin-bottom: 1rem;
-  }
-
-  & .button {
-    width: 200px;
-    margin-top: 2rem;
-    margin-left: auto;
-    margin-right: auto;
-    padding: 1rem;
-    background-color: #fff;
-    border-radius: 18px;
-    border: none;
-    box-shadow: 0 10px 15px rgba(0,0,0,0.1);
-    font-family: 'Euclid Triangle', 'Open Sans', sans-serif;
-  }
-
-  & .button:not(:disabled):hover {
-      cursor: pointer;
-  }
-
-  &.collapsed {
-      transform: translate(-100%);
-      box-shadow: 0 24px 15px rgba(0,0,0,0);
-  }
-
-  h1 {
-      line-height: 1.3;
-      color: black;
-      font-size: 1.6em;
-      margin: 0;
-  }
-
-  & > svg {
-      display: block;
-  }
-
-  &.text {
-      color: $black;
-  }
-
-  .title {
-      font-size: 1.3rem;
-      font-family: $headline-font;
-  }
-
-  @include respond-to('medium-up') {
-      height: 100%; /* Should only be set on desktop */
-  }
-
-  @include respond-to('small') {
-      padding-bottom: 0;
-      position: relative;
-      width: 100%;
-  }
-}
