+++ src/scss/styles.scss
@@ -2,7 +2,7 @@
 @import "mixins";
 @import "base";
 
-@import "layout/data-center-compare";
+@import "layout/data-center-compare-panel";
 @import "layout/header";
 @import "layout/footer";
 @import "layout/left-panel";
